Added: Añadido opcion de crear una nueva lista(sweetatalert2) Añadir guardar con boton en el .json tras cualquier accion. To do: Acabar UI Añadir al cerrar cada pop up(Cada lista) autoguardado(llamar a la funcion saveList(list)). Ej Swal.fire({}).then((result) => {saveList(list)});

+++ frontend/src/App.tsx
@@ -1,6 +1,3 @@
-// import { useSelector, useDispatch } from 'react-redux';
-// import { RootState, AppDispatch } from './redux/store';
-// import { loadData } from './redux/ToDoListSlice';
 import { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { useAppDispatch } from './redux/store';
@@ -11,13 +8,12 @@ import {
   deleteTask,
   addNewTask,
   deleteList,
+  createList,
+  saveLists,
 } from './redux/ToDoListSlice';
 import Swal from 'sweetalert2';
 
 export default function App() {
-  // const count = useSelector((state: RootState=> state.counter.value);
-  // const dispatch = useDispatch<AppDispatch>();
-  // const [list, setList] = useState<List[]>([]);
   const dispatch = useAppDispatch();
   const list = useSelector((state: RootState) => state.toDoList.Lists);
 
@@ -25,11 +21,31 @@ export default function App() {
     dispatch(fetchLists());
   }, []);
 
-  const handleToggleTaskFinished = (listId: number, taskId: number) => {
-    dispatch(toggleTaskFinished({ listId, taskId }));
+  const handleToggleTaskFinished = async (listId: number, taskId: number) => {
+    try {
+      // Aquí puedes llamar a tu acción para agregar la nueva tarea a la lista correspondiente
+      dispatch(toggleTaskFinished({ listId, taskId }));
+    } catch (error) {
+      console.error('Error al cambiar el estado de la tarea:', error);
+    }
   };
-  const deleteTaskFunc = (listId: number, taskId: number) => {
-    dispatch(deleteTask({ listId, taskId }));
+  const deleteTaskFunc = async (listId: number, taskId: number) => {
+    const { isConfirmed } = await Swal.fire({
+      title: 'Estas seguro de que desea eliminar la tarea',
+      showCancelButton: true,
+      confirmButtonText: 'Eliminar',
+      cancelButtonText: 'Cancelar',
+    });
+
+    if (isConfirmed) {
+      // Aquí puedes llamar a tu acción para agregar la nueva tarea a la lista correspondiente
+      try {
+        // Aquí puedes llamar a tu acción para agregar la nueva tarea a la lista correspondiente
+        dispatch(deleteTask({ listId, taskId }));
+      } catch (error) {
+        console.error('Error al eliminar la tarea:', error);
+      }
+    }
   };
   const addTaskFunc = async (listId: number, listTitle: string) => {
     const { value, isConfirmed } = await Swal.fire({
@@ -51,7 +67,12 @@ export default function App() {
 
     if (isConfirmed && value) {
       // Aquí puedes llamar a tu acción para agregar la nueva tarea a la lista correspondiente
-      dispatch(addNewTask({ listId, taskText: value }));
+      try {
+        // Aquí puedes llamar a tu acción para agregar la nueva tarea a la lista correspondiente
+        dispatch(addNewTask({ listId, taskText: value }));
+      } catch (error) {
+        console.error('Error al añadir la tarea:', error);
+      }
     }
   };
   const deleteListFunc = async (listId: number, listTitle: string) => {
@@ -78,7 +99,45 @@ export default function App() {
 
     if (isConfirmed && value) {
       // Aquí puedes llamar a tu acción para agregar la nueva tarea a la lista correspondiente
-      dispatch(deleteList({ listId }));
+      try {
+        // Aquí puedes llamar a tu acción para agregar la nueva tarea a la lista correspondiente
+        dispatch(deleteList({ listId }));
+      } catch (error) {
+        console.error('Error al eliminar la lista', error);
+      }
+    }
+  };
+  const createListFunc = async () => {
+    const { value, isConfirmed } = await Swal.fire({
+      title: 'Crear nueva lista',
+      input: 'text',
+      text: 'Introduce el nombre de la lista',
+      inputPlaceholder: 'Nombre de la lista',
+      showCancelButton: true,
+      confirmButtonText: 'Crear',
+      cancelButtonText: 'Cancelar',
+      inputValidator: (value) => {
+        if (value === '') {
+          return 'La tabla tiene que tener un titulo';
+        }
+      },
+    });
+
+    if (isConfirmed && value) {
+      try {
+        // Aquí puedes llamar a tu acción para agregar la nueva tarea a la lista correspondiente
+        dispatch(createList({ listTitle: value }));
+      } catch (error) {
+        console.error('Error al crear la lista:', error);
+      }
+    }
+  };
+  const saveDataFunc = async () => {
+    try {
+      // Aquí puedes llamar a tu acción para agregar la nueva tarea a la lista correspondiente
+      dispatch(saveLists(list));
+    } catch (error) {
+      console.error('Error al guardar la lista:', error);
     }
   };
 
@@ -87,7 +146,16 @@ export default function App() {
       <h1 className="text-3xl font-extrabold text-center text-gray-800 mb-8">
         Lista de Tareas
       </h1>
-
+      <button
+        onClick={() => createListFunc()}
+        className="px-4 py-2 bg-orange-300 text-white rounded-lg hover:bg-orange-500 transition duration-200">
+        Añadir Lista
+      </button>
+      <button
+        onClick={() => saveDataFunc()}
+        className="px-4 py-2 bg-gray-400 text-white rounded-lg hover:bg-gray-700 transition duration-200">
+        Guardar Lista
+      </button>
       <ul className="space-y-6">
         {list.map((item) => (
           <li
@@ -110,13 +178,11 @@ export default function App() {
                   Añadir Tarea
                 </button>
 
-                {item.tasks.length === 0 && (
-                  <button
-                    onClick={() => deleteListFunc(item.id, item.title)}
-                    className="px-4 py-2 bg-red-800 text-white rounded-lg hover:bg-red-900 transition duration-200">
-                    Eliminar Lista
-                  </button>
-                )}
+                <button
+                  onClick={() => deleteListFunc(item.id, item.title)}
+                  className="px-4 py-2 bg-red-800 text-white rounded-lg hover:bg-red-900 transition duration-200">
+                  Eliminar Lista
+                </button>
               </div>
             </div>
 
@@ -125,19 +191,22 @@ export default function App() {
                 <li
                   key={task.id}
                   className="p-4 rounded-lg bg-white shadow-md hover:shadow-lg transition duration-200">
-                  <div className="flex justify-between items-center">
+                  <div
+                    className="flex justify-between items-center cursor-pointer"
+                    onClick={() => handleToggleTaskFinished(item.id, task.id)} // Se activa al hacer clic en el div
+                  >
                     <p
-                      className={`cursor-pointer text-gray-800 ${
+                      className={`text-gray-800 ${
                         task.isFinished ? 'line-through text-gray-500' : ''
-                      }`}
-                      onClick={() =>
-                        handleToggleTaskFinished(item.id, task.id)
-                      }>
+                      }`}>
                       {task.text}
                     </p>
 
                     <button
-                      onClick={() => deleteTaskFunc(item.id, task.id)}
+                      onClick={(event) => {
+                        event.stopPropagation(); // Evita que el evento llegue al div
+                        deleteTaskFunc(item.id, task.id);
+                      }}
                       className="px-4 py-2 bg-red-500 text-white rounded-lg hover:bg-red-600 transition duration-200">
                       Eliminar
                     </button>
